+++ com.unity.multiplayer.mlapi/Tests/Runtime/NetworkSceneManagerCallbacksTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 271404786bd294e4e9f688dc1daaa883
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
