+++ com.unity.multiplayer.mlapi/Tests/Runtime/NetworkSceneManagerTests.cs
@@ -32,71 +32,6 @@ namespace MLAPI.RuntimeTests
             Assert.IsTrue(threwException);
         }
 
-        [UnityTest]
-        public IEnumerator TestOnClientLoadedScene()
-        {
-            SetupNetworkManagerForClientLoadedCallbackTests();
-            var networkManager = NetworkManagerHelper.NetworkManagerObject;
-            
-            bool callbackReceived = false;
-            
-            networkManager.SceneManager.OnClientLoadedScene += (progress, clientId) =>
-            {
-                Debug.Log("OnClientLoadedScene invoked on the host, all good.");
-                callbackReceived = true;
-            }; 
-            
-            networkManager.SceneManager.SwitchScene("SceneWeAreSwitchingTo");
-
-            var startTime = DateTime.UtcNow;
-            do
-            {
-                yield return null;
-            }
-            while ((DateTime.UtcNow - startTime).TotalSeconds < networkManager.NetworkConfig.LoadSceneTimeOut && !callbackReceived);
-            
-            Assert.IsTrue(callbackReceived);
-        }
-        
-        [UnityTest]
-        public IEnumerator TestOnAllClientsLoadedScene()
-        {
-            SetupNetworkManagerForClientLoadedCallbackTests();
-            var networkManager = NetworkManagerHelper.NetworkManagerObject;
-            
-            bool callbackReceived = false;
-            
-            networkManager.SceneManager.OnAllClientsLoadedScene += (progress, timedOut) =>
-            {
-                Debug.Log("OnAllClientsLoadedScene invoked on the host, all good.");
-                callbackReceived = true;
-            }; 
-            
-            networkManager.SceneManager.SwitchScene("SceneWeAreSwitchingTo");
-
-            var startTime = DateTime.UtcNow;
-            do
-            {
-                yield return null;
-            }
-            while ((DateTime.UtcNow - startTime).TotalSeconds < networkManager.NetworkConfig.LoadSceneTimeOut && !callbackReceived);
-            
-            Assert.IsTrue(callbackReceived);
-        }
-
-        private void SetupNetworkManagerForClientLoadedCallbackTests()
-        {
-            Assert.IsTrue(NetworkManagerHelper.StartNetworkManager(out NetworkManager networkManager, NetworkManagerHelper.NetworkManagerOperatingMode.None));
-            Debug.Log("We have stopped the NetworkManager to do changes to NetworkConfig.");
-
-            networkManager.NetworkConfig.EnableSceneManagement = true;
-            networkManager.NetworkConfig.RegisteredScenes.Add("SceneWeAreSwitchingTo");
-            networkManager.NetworkConfig.LoadSceneTimeOut = 5;
-
-            networkManager.StartHost();
-            Debug.Log("Host started");
-        }
-
         [SetUp]
         public void Setup()
         {
